backend/requirements.txt
Django==4.2.20
djangorestframework==3.15.2
pillow==11.0.0
uvicorn==0.32.0
channels==4.1.0
channels-redis==4.2.0
whitenoise==6.8.2
django-redis==5.4.0
psycopg==3.1.18
django-environ==0.9.0
django-ckeditor==6.7.2
celery==5.4.0
django-celery-results==2.5.1